made Example.groovy repeatable

+++ src/example/groovy/Example.groovy
@@ -1,7 +1,7 @@
 import com.curiousattemptbunny.grouch.*
 
 grouch = new Grouch('http://localhost:5984')
-grouch.mydb.delete()
+if (grouch.mydb.exists()) grouch.mydb.delete()
 grouch.mydb.create()
 grouch.mydb << [_id: 'mydoc', aField: 'aValue' ]
 assert grouch.mydb.mydoc.aField == 'aValue'
